url/host-parser-reduced.html

<!doctype html>
<meta charset=utf-8>
<title>Testing the host parser</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
let setup = async_test("Loading data…")
setup.step(() => {
  const tests = { "asciihost": null, "IdnaTest": null }
  let testNumber = Object.keys(tests).length
  Object.keys(tests).forEach((val) => {
    let request = new XMLHttpRequest()
    request.open("GET", val + ".json")
    request.send()
    request.responseType = "json"
    request.onload = setup.step_func(() => {
      tests[val] = request.response
      testNumber--
      if(testNumber === 0) {
        Object.keys(tests).forEach((val) => {
          runHostTests(val, tests[val])
        })
        setup.done()
      }
    })
  })
})

function runHostTests(source, hostTests) {
  for(var i = 0, l = hostTests.length; i < l; i++) {
    let hostTest = hostTests[i]
    if (typeof hostTest === "string") {
      continue // skip comments
    }
    let outputInput = hostTest.input
    //if(hostTest.input.length < 4) {
    //  outputInput = encodeURI(hostTest.input)
    //}

    test(() => {
      const url = new URL("https://x/x")
      url.host = hostTest.input
      if(hostTest.output !== null) {
        assert_equals(url.host, hostTest.output)
      } else {
        assert_equals(url.host, "x")
      }
    }, "Parse host using URL's host setter: " + outputInput + " (from: " + source + ")")
  }
}
</script>
